refactor: Extract independent test helper functions to separate files
