+++ tests/end_to_end_cases/mod.rs
@@ -1,0 +1,2 @@
+pub mod grpc_api;
+pub mod read_api;
